+++ R/seasonal_burden_levels.R
@@ -62,7 +62,7 @@
 #'                          to = end_date,
 #'                          by = "week")
 #'
-#' tsd_data <- tsd(
+#' tsd_data <- to_time_series(
 #'   cases = c(season_1, season_2),
 #'   time = as.Date(weekly_dates)
 #' )