+++ NEWS.md
@@ -10,6 +10,8 @@
 
 * Added new argument `use_offset` to `seasonal_onset()` that allows the user to add the background population to adjust the growth rate estimations (#83).
 
+* Added new feature to estimate multiple waves in `combined_seasonal_output` (#77).
+
 ## Improvements
 
 * Observations are now divided into `cases` and `incidence`, which is implemented into all functions in the package. Cases are used as default, but if the user additionally inputs `population` the output will be incidence (#84).