+++ NEWS.md
@@ -1,4 +1,4 @@
-# aedseo (development version)
+# aedseo 1.0.0
 
 ## Deprecations